+++ macrel/output.py
@@ -142,3 +142,29 @@ readme_output_reads_mode = f"""{header}
 
 {footer}
 """
+
+readme_output_ampsphere_mode = f"""{header}
+
+For results for this mode, please see more information and cite the AMPSphere manuscript:
+
+> **Discovery of antimicrobial peptides in the global microbiome with machine
+> learning** by Célio Dias Santos Júnior, Marcelo D.T. Torres, Yiqian Duan,
+> Álvaro Rodríguez del Río, Thomas S.B. Schmidt, Hui Chong, Anthony Fullam,
+> Michael Kuhn, Chengkai Zhu, Amy Houseman, Jelena Somborski, Anna Vines,
+> Xing-Ming Zhao, Peer Bork, Jaime Huerta-Cepas, Cesar de la Fuente-Nunez, Luis
+> Pedro Coelho in _Cell_
+> [doi:10.1016/j.cell.2024.05.013](https://doi.org/10.1016/j.cell.2024.05.013)
+
+
+## Outputs for `ampsphere` mode
+
+There are three modes:
+
+- `exact`: exact match only
+- `mmseqs`: MMSeqs2 match
+- `hmm`: HMMER match
+
+The exact output format will depend on the mode as it will return results from each tool.
+
+{footer}
+"""